+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog
 ### New Features
 
 * add `contributions/count[/density]/groupBy/boundary` endpoints to fetch contribution counts/densities grouped by boundary ([#217])
+* add an error message for requests without defined parameters ([#227])
 
 ### Bug Fixes
 
@@ -13,9 +14,9 @@ Changelog
 * fix a regression causing `…/groupBy/boundary` to crash when used with a geometry type filter using the (deprecated) types/keys/values parameter syntax ([#230])
 
 [#217]: https://github.com/GIScience/ohsome-api/issues/217
+[#227]: https://github.com/GIScience/ohsome-api/issues/227
 [#228]: https://github.com/GIScience/ohsome-api/pull/228
 [#230]: https://github.com/GIScience/ohsome-api/pull/230
-
 
 ## 1.6.0